producer endpoint with pageable list all and exception handler

+++ src/main/java/com/kasemodel/goldenraspberryaward/application/impl/ProducerServiceImpl.java
@@ -2,17 +2,21 @@ package com.kasemodel.goldenraspberryaward.application.impl;
 
 import com.kasemodel.goldenraspberryaward.application.ProducerService;
 import com.kasemodel.goldenraspberryaward.infra.persistence.entity.Producer;
-import com.kasemodel.goldenraspberryaward.infra.persistence.exception.ProducerAlreadyExistsException;
-import com.kasemodel.goldenraspberryaward.infra.persistence.exception.ProducerNotFoundException;
+import com.kasemodel.goldenraspberryaward.infra.persistence.exception.producer.ProducerAlreadyExistsException;
+import com.kasemodel.goldenraspberryaward.infra.persistence.exception.producer.ProducerNameCannotBeEmptyException;
+import com.kasemodel.goldenraspberryaward.infra.persistence.exception.producer.ProducerNotFoundException;
 import com.kasemodel.goldenraspberryaward.infra.persistence.repository.ProducerRepository;
+import com.kasemodel.goldenraspberryaward.interfaces.rest.model.CreateByNameRequest;
 import com.kasemodel.goldenraspberryaward.interfaces.rest.model.ProducerResponse;
 import jakarta.persistence.Tuple;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -20,7 +24,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 @Service
 @Slf4j
-public class ProducerServiceImpl implements ProducerService {
+public class ProducerServiceImpl extends PageableServiceImpl implements ProducerService {
 	private final ProducerRepository repo;
 
 	@Override
@@ -42,16 +46,14 @@ public class ProducerServiceImpl implements ProducerService {
 	}
 
 	@Override
-	public Producer validateAndCreate(final String name) throws ProducerAlreadyExistsException {
-		if (repo.existsByName(name))
-			throw new ProducerAlreadyExistsException(name);
-		return create(name);
+	public Producer validateAndCreate(final CreateByNameRequest producer) throws ProducerAlreadyExistsException {
+		validate(producer);
+		return create(producer.name());
 	}
 
 	@Override
-	public List<Producer> findAll() {
-		return repo.findAllByDeletedAtIsNull()
-			.orElseGet(Collections::emptyList);
+	public Page<Producer> findAll(final PageRequest pageable) {
+		return repo.findAllByDeletedAtIsNull(pageable);
 	}
 
 	@Override
@@ -66,6 +68,8 @@ public class ProducerServiceImpl implements ProducerService {
 	@Override
 	@Transactional
 	public void delete(UUID externalId) {
+		if (!repo.existsByExternalId(externalId))
+			throw new ProducerNotFoundException(externalId);
 		repo.deleteByExternalId(externalId);
 	}
 
@@ -78,5 +82,12 @@ public class ProducerServiceImpl implements ProducerService {
 	@Transactional
 	private Producer create(final String name) {
 		return repo.save(Producer.of(name));
+	}
+
+	private void validate(final CreateByNameRequest producer) {
+		if (producer == null || StringUtils.isBlank(producer.name()))
+			throw new ProducerNameCannotBeEmptyException();
+		if (repo.existsByName(producer.name()))
+			throw new ProducerAlreadyExistsException(producer.name());
 	}
 }
